[TS Migration] PR B: Convert SharedStyles and serviceWorker to TypeScript
- SharedStyles.js → SharedStyles.ts: add SxProps<Theme> types to all MUI sx exports
- serviceWorker.js → serviceWorker.ts: add ServiceWorkerConfig interface, type all
  function signatures, replace process.env with import.meta.env (Vite equivalents)
- tsconfig.json: add vite/client to types for import.meta.env support

Verification: npm run build ✓  tsc --noEmit ✓  npm test ✓
