--- src/common/styles/SharedStyles.js
@@ -1,20 +1,0 @@
-export const headerTextContainerSx = { mt: 3, mb: 3 };
-
-export const paperTextContainerSx = { ml: 3, mr: 3, mb: 2, background: '#324a54' };
-
-export const headerTextSx = {
-    fontSize: '2.6em',
-    fontWeight: 'bold',
-    ml: 3,
-    mb: 2,
-    fontFamily: 'Quicksand, sans-serif'
-};
-
-export const paperTextSx = {
-    fontSize: '1.3em',
-    color: 'white',
-    fontFamily: 'Roboto, sans-serif',
-    mt: 1,
-    p: 3,
-    lineHeight: '45px'
-};
