+++ tsconfig.json
@@ -15,6 +15,7 @@
     "forceConsistentCasingInFileNames": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
+    "types": ["vite/client"],
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true
   },
